Logging fix's
git-svn-id: https://arducopter.googlecode.com/svn/trunk@1888 f9c3cf11-9bcb-44bc-f272-b75c42450872

+++ ArduCopterMega/ArduCopterMega.pde
@@ -611,7 +611,7 @@ void medium_loop()
 			medium_loopCounter++;
 
 			if (g.log_bitmask & MASK_LOG_ATTITUDE_MED && (g.log_bitmask & MASK_LOG_ATTITUDE_FAST == 0))
-				Log_Write_Attitude((int)dcm.roll_sensor, (int)dcm.pitch_sensor, (int)dcm.yaw_sensor);
+				Log_Write_Attitude((int)dcm.roll_sensor, (int)dcm.pitch_sensor, (uint16_t)dcm.yaw_sensor);
 
 			#if HIL_MODE != HIL_MODE_ATTITUDE
 			if (g.log_bitmask & MASK_LOG_CTUN)
